requirements.txt
langchain-core
langchain-groq
langchain-ollama
python-dotenv
pyyaml 
